Update nginxpla config for better URL categorization and logging
This commit updates the nginxpla configuration file to improve URL categorization and enhance the logging mechanism. We updated URL categories to make them more specific, increasing the recognition of pages such as '/sale', '/shop', '/search', and more. This provides more accurate data for analysis and helps to differentiate various site interactions. Furthermore, we extended reporting parameters specifically for browser-related data which assist in analyzing user behaviour. The change will not affect any existing functionality but it enhances the specificity and depth of our logging capability.